requirements.txt
fastapi
uvicorn
python-dotenv
openai
pymongo
python-jose[cryptography]
passlib
argon2-cffi
google-generativeai
